# server/openapi.yaml
openapi: 3.0.0
info:
  title: n.eko REST API
  description: Next Gen Renderer.
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: "1.0.0"
servers:
  - description: Local server
    url: http://localhost:3000
...
paths:
...
  /api/profile:
    post:
      summary: update current profile without syncing it with member profile (experimental)
      operationId: profile
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberProfile'
        required: true
...
  /api/room/control:
    get:
      tags:
        - room
      summary: get control status
      operationId: controlStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/request:
    post:
      tags:
        - room
      summary: request control
      operationId: controlRequest
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/release:
    post:
      tags:
        - room
      summary: release control
      operationId: controlRelease
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/take:
    post:
      tags:
        - room
      summary: take control
      operationId: controlTake
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/give/{sessionId}:
    post:
      tags:
        - room
      summary: give control
      operationId: controlGive
      parameters:
        - in: path
          name: sessionId
          description: session ID
          required: true
          schema:
            type: string
      responses:
        '204':
          description: OK
        '400':
          description: Target session is not allowed to host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/room/control/reset:
    post:
      tags:
        - room
      summary: reset control
      operationId: controlReset
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/members:
    get:
      tags:
        - members
      summary: list of members
      operationId: membersList
      parameters:
        - in: query
          name: limit
          schema:
            type: number
        - in: query
          name: offset
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - members
      summary: create new member
      operationId: membersCreate
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Member with chosen ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
        required: true
  /api/members/{memberId}:
    get:
      tags:
        - members
      summary: get member's profile
      operationId: membersGetProfile
      parameters:
        - in: path
          name: memberId
          description: member identifier
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - members
      summary: update member's profile
      operationId: membersUpdateProfile
      parameters:
        - in: path
          name: memberId
          description: member identifier
          required: true
          schema:
            type: string
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberProfile'
        required: true
    delete:
      tags:
        - members
      summary: remove member
      operationId: membersRemove
      parameters:
        - in: path
          name: memberId
          description: member identifier
          required: true
          schema:
            type: string
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/members/{memberId}/password:
    post:
      tags:
        - members
      summary: update member's password
      operationId: membersUpdatePassword
      parameters:
        - in: path
          name: memberId
          description: member identifier
          required: true
          schema:
            type: string
      responses:
        '204':
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberPassword'
        required: true
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: NEKO_SESSION
    BearerAuth:
      type: http
      scheme: bearer
    TokenAuth:
      type: apiKey
      in: query
      name: token
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
...
    ControlStatus:
      type: object
      properties:
        has_host:
          type: boolean
        host_id:
          type: string
...
    MemberProfile:
      type: object
      properties:
        name:
          type: string
        is_admin:
          type: boolean
        can_login:
          type: boolean
        can_connect:
          type: boolean
        can_watch:
          type: boolean
        can_host:
          type: boolean
        can_share_media:
          type: boolean
        can_access_clipboard:
          type: boolean
        sends_inactive_cursor:
          type: boolean
        can_see_inactive_cursors:
          type: boolean
        plugins:
          type: object
          additionalProperties: true
    MemberData:
      properties:
        id:
          type: string
        profile:
          $ref: '#/components/schemas/MemberProfile'
    MemberCreate:
      properties:
        username:
          type: string
        password:
          type: string
        profile:
          $ref: '#/components/schemas/MemberProfile'
    MemberPassword:
      properties:
        password:
          type: string
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - TokenAuth: []